receiver/requirements.txt
pyrtlsdr==0.2.93
